src: fix asm-verify cross-build compile regressions (BakedString/AboutScreen/ItemManager)
The asm-verify cross-build (GCC 4.4.1, -D__bada__ -DFN_ASM_VERIFY_CROSS,
-fno-rtti) caught three portability regressions from this branch that the
lenient host MSVC build missed:

- BakedString.h (360a467): forward-declared QUADCUSTOMVERTEX inside namespace
  Mortar, shadowing the real global ::QUADCUSTOMVERTEX so m_pPageVertices
  resolved to an incomplete Mortar:: type. Moved the forward-decl to global
  scope; .cpp now sees the complete type.
- AboutScreen.h (da705d4): the offsetof() layout static_asserts referenced
  PRIVATE members from an external struct, which GCC 4.4.1 rejects. Make the
  layout members public under #if defined(__bada__) (host keeps them private)
  so the asserts still FIRE in the cross-build (the 32-bit binary ABI) instead
  of being disabled there.
- ItemManager.cpp (coin reroute): SetAttribute(const char*, int32_t&) is
  ambiguous in the cross-toolchain (int32_t == long); added (int) casts at the
  3 call sites.

Host build clean; cross-build no longer errors on these TUs. Remaining
cross-build failure is the pre-existing shared_ptr/typeid-with-fno-rtti issue
(Geometry.h #include <memory>, present since base 01f64df), unrelated to this
branch.

+++ src/game/ItemManager.cpp
@@ -350,9 +350,9 @@ void ItemManager::SaveItemInfo() {
     // Coin balance lives in game_work (+0x20/+0x24/+0x28), not FruitSaveData.
     tinyxml2::XMLElement* root = doc.NewElement("item_save_file");  // 0x1b9e4d
     root->SetAttribute("version", "1.0");  // 0x1b9e5c / 0x1b9e64
-    root->SetAttribute("coins",           game_work.m_CoinsBalance);
-    root->SetAttribute("coinsTotal",      game_work.m_CoinsTotalEarned);
-    root->SetAttribute("levelStartCoins", game_work.m_CoinsAtGameStart);
+    root->SetAttribute("coins",           (int)game_work.m_CoinsBalance);
+    root->SetAttribute("coinsTotal",      (int)game_work.m_CoinsTotalEarned);
+    root->SetAttribute("levelStartCoins", (int)game_work.m_CoinsAtGameStart);
 
     // <boughtItems> section: all items with m_Cost < 0 (purchased)
     tinyxml2::XMLElement* bought = doc.NewElement("boughtItems");  // 0x1b9e89
